refractor: add seperate transformer
